bugfix: multi-file matfiles loading incorrect number of timepoints
Due to this pre_processing routine being split into 2 separate
functions, we now need to load the dimensions of each file rather than
relying on variables still in memory. During the array initialization,
the variable T*length was used for the total number of timestamps, but
the actual extraction of data used just T. So our final array was
overwriting on itself rather than appending to the end of the previous
array.